// types/server/contracts.d.ts
  interface QueryRes<T> {
    query_result: T;
  }

  /** _shared */
  type Vote = "yes" | "no";

  type EmbeddedWasmMsg = {
    wasm: {
      execute: {
        contract_addr: string;
        funds: Coin.Data[];
        msg: string; //base64 endocoded msg object
      };
    };
  };

  type EmbeddedBankMsg = {
    bank: {
      send: {
        amount: Coin.Data[];
        to_address: string;
      };
    };
  };
// ...
  type Member = {
    addr: string;
    weight: number;
  };

  type MemberRes = {
    members: Member[];
  };

  type ProposalsRes = {
    proposals: Proposal[];
  };

  type ProposalStatus = "pending" | "open" | "rejected" | "passed" | "executed";

  type Proposal = {
    id: number; //1
    title: string; //"this prpposal rocks"
    description: string; //"this is a description"
    meta?: string; //JSON string that contains preview metadata
    msgs: (EmbeddedWasmMsg | EmbeddedBankMsg)[];
    status: ProposalStatus;
    expires: { at_height: number };
    threshold: {
      //others absolute account, threshold quourum
      absolute_percentage: {
        percentage: string; //"0.5"
        total_weight: number; //2
      };
    };
  };

  type CW3Config = {
    group_addr: string; //"terra123abc.."
    threshold: {
      absolute_percentage: {
        percentage: string; //"0.5"
      };
    };
    max_voting_period: {
      height: number; //1000
    };
    isPlacholder?: true;
    //...future needed attr
  };

  type AdminVoteInfo = {
    voter: string; //"terra123abc.."
    vote: Vote;
    weight: number; //1
  };